feat: Added -j/--json output option to cli-validator
This revision adds a new --json option to the cli-validator that prints the internal
result data in json format to stdout.  It does additionally include the line number
as a 'line' field in the JSON (this is not in the internal json data).  Automated
test validates that json is received.

Closes #43