feat: Added --config option, adding support for explicit config file
This feature adds a -c/--config option which is a path to a config file (.validaterc).
This allows usage of configuration that are not dependent on the .validaterc being in
the same folder (or a parent) of the current working directory.

Note that the automated test added for this doesn't run currently.  For the test I am
changing values for the defaults (a warning to error and other warning to off) so that
the reported errors in the justWarn.yaml file are changed from stock behavior.  Manually
executing the commands from the command line works as expected, but it seems that in the
context of the cli tests that the config seems to be totally ignored and the default is
all that is used.  I have not been able to figure this out yet... Any pointers as to why
this is and how to fix it are appreciated.

Closes #49