fix: alllow absolute filepaths to be passed to --config
It turns out that the findUp method cannot handle a full path to a file as input. When this
is passed in, it redudantly or completely incorrectly prepends the current working directory
to the front of the returned path.  Work around is to split any path information off of the
input path and pass the path portion into the method as opt.cwd.