feat(plugin): adds k8s_camel_case validator for k8s API conventions
The Kubernetes API conventions use lower_camel_case except for acronyms which are expected to be all
the same case.  When an acronym is not at the beginning of the string it should be capitalized.
This change adds a new validator called k8s_camel_case that relaxes lower_camel_case to allow this.

https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#naming-conventions

fix #105