fix: bundle original spec to account for multi-file API definitions
We do some validations on the "resolved spec" or the full API definition with all of its
references resolved into one large document. This takes multi-file specs into account and
resolves all references, even to other files. However, some validations are performed only
on the original spec as it provides some benefits, such as gathering all schemas without
worrying about duplicates. The problem with our current approach is that, if the spec is
split into multiple files, we only validate the entry file and none of the others.

This adds a quick step to "bundle" all of the files together into one "file" that only has
internal references. This way, the unresolved "js spec" will represent the entire definition.