fix: properly handle circular regressions to prevent call stack exceptions (#253)
A regression was recently introduced to the validator in the circular references
utility that prevented such references from being caught and sanitized. This resulted
in "call stack exceeded" expections for certain patterns using circular references.

The regression came in the form of checking for only "plain" objects, rather than
all "objects" (which include arrays). This happened in a recent community PR and I
didn't realize the consequences when I reviewed and merged. I went back and double
checked every instance of this change in that PR and believe this is the only one
that is problematic. I fixed it and added a new test case that would've caught this
regression had it been there.