feat: upgrade to spectral 6 and convert to monorepo (#349)
This commit refactors the project into a monorepo and separates the project
into two NPM packages. One is the existing ibm-openapi-validator package and
one is the new package for our custom Spectral ruleset. As part of this change,
we upgraded to Spectral 6 which contained a number of breaking changes.

There are some incompatible changes for this tool (not many) as a result. These
are described in the migration guide. Since a number of these files were being
moved around already, I took the opportunity to rename a bunch of files to match
our kebab-case convention of file naming.