feat: enable spectral overrides
This commit updates the format in which we provide the document data to spectral,
which in turns allows end users to take advantage of spectral "overrides". This
feature allows users to ignore specific parts of their document if they need,
rather than turning off entire rules.

In doing this, we give spectral greater control over the validations it returns.
This allows us to take advantage of another spectral feature - automatically
resolving component paths. Before, if a schema contained a problem and two references
pointed to it, some rules would return two validations pointing to the reference
instances. Now, the validation returns one validation, pointing to the base schema.
This will result in some output changes for some users but it should be a positive
change overall.

This commit also adds a customized verson of spectral's deduplication logic to prevent
the resolved component paths from being duplicated.