feat: add new spectral `discriminator` rule and remove old rule (#367)
ruleset: add 'discriminator' rule to ruleset

This rule verifies that the 'propertyName' value of the discriminator
is present in each subschema.

validator: remove broken discriminator rule in favor of new spectral rule

This rule was buggy and not previously configurable. The same rule with
identical (albeit fixed) behavior is now available in our Spectral ruleset
under the name 'discriminator' and is configurable.