feat: add new `property-inconsistent-name-and-type` rule and remove old rule
This removes the old `inconsistent_property_type` rule in favor of a new, Spectral
rule called `property-inconsistent-name-and-type`. This performs the same function -
it validates that all properties in an API definition of the same name use a consistent
type. This also updates the error message from what it was before so that it is more
clear.