refactor: move $ref check out of individual functions (#471)
When invoking a function on a rule with the resolved spec, it is prudent to
check for schemas defined by a $ref. In the resolved spec, schemas defined with
a $ref are the result of circular references being detected and it isn't
necessary to validate the schemas.

To avoid duplicating this check for each function and requiring manual
maintenance in all functions (many were missing the check), I moved it to the
validate functions that they all rely on to recursively check schemas.

The only potential downside here is that it would prevent us from writing a
function that recursively walks through the unresolved spec and that also
relies on the presence of $refs in the schemas. This feels extremely unlikely
to me so it seems a worthwhile optimization.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>