fix(string-boundary): improve handling of composed schemas
This commit modifies the "string-boundary" rule so that it
more accurately performs the checks for the various schema
attributes within composed schemas.
For example, if a string schema contains an allOf, then the
"pattern" field could be specified in any of the allOf
list element schemas instead of the main schema itself.
In contrast, for a string schema containing a oneOf or anyOf,
the "pattern" field would need to be specified in either
the main schema itself OR within EACH of the oneOf/anyOf
list element schemas.

Previously, the rule was not accurate enough with its
checks within composed schemas.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>