fix(response-status-codes): avoid false positives for 'create' operations (#483)
This commit fixes a problem with the 'response-status-codes'
validation rule.  Prior to this change, any POST request would be considered
a "create" operation.  Now, a POST request will be considered a
"create" operation only if the API also defines a corresponding
get/put/patch/delete path (i.e. a similar path but with a trailing
path param reference.
In any case, an operation whose operationId starts with "create"
will continue to be considered a "create" operation regardless of the
paths defined in the API.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>