feat(optional-request-body): add new spectral-style rule
This commit introduces a new spectral-style
'optional-request-body' rule, which will examine
each optional request body and it's associated schema
to see if it looks like the API author might have
intended to define the requestBody as required.
It is rare to have an optional requestBody with
a schema that contains required properties.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>