fix: fix broken pkg script and publish binaries to github
The pkg script was broken because of some clever usage of "rename" that wasn't
compatible with all devices. This commit removes the need for it by moving the
logic into a separate bash script with way more clarity.

This also updates semantic release to include the built platform binaries in
the "assets" section of the release on GitHub.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>