fix(property-case-convention): report correct path in errors
This commit modifies the "property-case-convention" rule
slightly so that when it reports a problem with the case
for a specific property, the offending property name will
be included in the error message, and the path associated with
the error will reflect the schema's "properties" field in
which the offending property is defined.
This subtle change will ensure that the spectral "unresolver"
will not be overzealous when "unresolving" the path.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>