refactor: move select utility functions into new package
Move out selected utilities and the "collections" module to a new folder,
which the new NPM package is created from. The ruleset package now depends
on this new utilities package.

This commit updates the ruleset package with the modified code to depend
on the new package and it sets up the new package.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>