refactor: remove validator module API (#539)
The Node module API for the validator is removed. The validator is available
as a CLI tool only. If use within a Node project is desired, use Spectral
within the JS code to run the IBM OpenAPI Ruleset.

BREAKING CHANGE: The Node interface of this tool is no longer supported, it is now a CLI tool only.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>