feat: add support for consistent config-file/cli options
This commit modifies the validator so that it supports
a consistent set of options via the command-line and
with a configuration file.

The configuration file can be specified with the -c/--config
option, and can be .js/.json/.yaml/.yml.

The cli options were updated to add support for
the -w/--warnings-limit and -i/--ignore options.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>