feat: unify all configuration options (#559)
This commit modifies the validator so that it supports a consistent set of
options via the command-line and with a configuration file.

The configuration file can be specified with the -c/--config option, and can be
.js/.json/.yaml/.yml.

The cli options are updated to add support for the -w/--warnings-limit and
-i/--ignore options.

BREAKING CHANGE: Support for the .thresholdrc and .validateignore files is removed. Use CLI options or the newly-supported configuration file.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>