refactor: apply config options to all output formats
The configuration options, such as "errors only" and "summary only", need to
apply to the JSON output, as well as the text output. This commit refactors
the logic so that the configuration options are honored during the actual
construction of the result object, so that it always represents the results
according to the configuration.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>