fix: include missing dependencies in ruleset package
The `chalk` dependency was missing from the ruleset package, which would prevent
users from using the ruleset as a standalone dependency. This commit adds it in.

Additionally, this commit makes the Spectral ruleset error handler a little
smarter. When there are validation errors in the ruleset, Spectral returns
an array of errors and we were losing the message, thus providing no info
to the user. This adds a check to extract the actual errors in these cases.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>