fix(ibm-define-required-properties): avoid false positives
Fixes: arf/planning-sdk-squad#3592

This commit fixes the ibm-define-required-properties rule
to avoid false positives.  Specifically the logic related
to composed models was updated to (hopefully) perform a more
accurate set of checks on the schemas in the API definition.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>

fix: edit logic to recursively account for nested composed schemas (#591)

and add a test case to demonstrate gap

Signed-off-by: Dustin Popp <dpopp07@gmail.com>