feat: add initial support for openapi 3.1 documents
This commit modifies the validator so that it can be
used to validate OpenAPI 3.1 documents.
This initial support does not yet include any
OpenAPI 3.1-specific rules in the IBM Validation Ruleset,
so the same rules used to validate OpenAPI 3.0.x documents
will be used to validate OpenAPI 3.1.x documents.
An exception to this is the set of rules inherited from
the Spectral OAS ruleset, since they do in fact implement
some OpenAPI 3.1-specific behavior within those rules.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>