fix(docker): publish container image upon successful release
The container was not being published because it was set up to push on a
tagged build, but we skip tagged builds (which I forgot about). This
updates the image publishing to push from semantic release's
"success" step, which will guarantee it gets pushed after the creation
of a new release during the main branch build.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>