fix(imports): use correct imports for lodash modules
The `lodash` packages supports installing individual modules via NPM but
we use enough modules that we install the whole package. However, we were
using the import style for individual modules, which didn't break because
the path in `node_modules` was still correct but caused warnings in
Spectral (when using custom YAML or JSON rulesets) because of the
incorrect names.

This resolves those warnings, which are annoying to users.

Resolves #534

Signed-off-by: Dustin Popp <dpopp07@gmail.com>