fix(build): run all release steps for validator build
The multi-sem-rel package we use merges the individual release configs
into the base project's config. I thought that it deep-merged the arrays
but it turns out that the validator package's `plugins` field was overriding
the whole `plugins` field in the root config and so we were never actually
going to release the package (because the NPM plugin would never be called).

This adds the rest of the plugins we call in the base config to the validator
config, so things should return to normal.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>