feat(new-rule): ibm-schema-naming-convention (#627)
Adds a new rule that enforces the schema naming guidelines in the API Handbook.

Specifically, the new rule checks for correctly named collection schemas, resource
collection element schemas, creation/replacement schemas, and patch schemas against
the name of the associated canonical schema.

There is one expection for creation/replacement schemas The Handbook states:

"There is an important exception to this guidance if the nature of an API is
to provide clients with full control over entire resources, with no
system-defined or immutable aspects. For such APIs, the canonical schema
itself SHOULD be used in requests to create or replace a resource."

The rule allows for the name of the canonical schema to be used for
resource-oriented POSTs and PUTs.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>
Co-authored-by: Phil Adams <phil_adams@us.ibm.com>