fix(ibm-schema-casing-convention): allow capitalized acronyms
It is a common convention to allow capitalized acronyms in schema names,
like `IAMCredentialsSecret`. Spectral's `casing` function does not provide
us with the flexibility to support this outright, so this PR adjusts the
schema casing convention check to use a regular expression instead (the one
used in the old, pre-Spectral validator).

Signed-off-by: Dustin Popp <dpopp07@gmail.com>