fix(api-symmetry): handle nested reference schemas
If reference schemas were nested in composed properties on the canonical
schema, the code would not resolve them the way it would for non-nested
reference schemas.

This moves the resolving code to the constraint checker, so that every
schema and applicator sub-schema will be resolved if it is a reference.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>