feat: display actual ruleset version with --version option
Before this commit, the --version option would print the version of the
validator tool and the version of the IBM Cloud OpenAPI Ruleset that
is run by default if the user doesn't provide their own ruleset.

However, if the user did use their own, locally installed version of our
ruleset, the version included in the output of the tool would not
actually reflect the version being used. Additionally, users could
install new versions of the validator tool but if they didn't update
their locally installed instance of the ruleset, they would not
get the updated behavior of the rules (which are often the reasons
for bumping the version of the tool).

This commit adds logic to print the actually used (whether the
default or locally installed) ruleset version and additionally will
print a warning if the user has upgraded the version of their tool
without upgrading the version of their locally installed ruleset.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>
Co-authored-by: Phil Adams <phil_adams@us.ibm.com>