fix(ibm-resource-response-consistency): expand support to bulk operations
This commit adds support for recognizing bulk operations and recommending the
collection schema be returned instead of the canonical schema. It also:
- Updates the message to be more clear
- Updates the path to be more aligned with the operation
- Adds an "info" log pointing to the specific operation in case it is unclear

Previously, the rule would flag a bulk operation that didn't return the collection
schema but it did so unintentionally (and with a bad heuristic) and the message
would indicate that the canonical schema should be returned, which was false (often,
the canonical schema *was* being returned).

Also, the warning would be tied to a specific schema and it was difficult
to parse the debug logs to determine which operation was responsible for the warning.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>