fix: adjust header parameter casing convention check (#648)
This allows us to support standard header patterns that would previously
be flagged by the validator. The `casing` function options provided by
Spectral are not sophsiticated enough to support the complexity of
header parameter conventions (which use kebab-separated pascal case with
capitalized abbreviations).

The new regular expression is a little more relaxed than the old convention
but it also more accurate.

I made this change in a configuration-compatible way by expanding the config
support to accept function arguments for either the casing or pattern built-in
Spectral functions. This is slightly hacky but gives both us and users more
control without changing any existing behavior. We might want to make this change
for our other casing convention checks but for now, this is sufficient to solve
our issues.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>
Co-authored-by: Phil Adams <phil_adams@us.ibm.com>