fix(build): build executables with correct dependency versions
This commit modifies the create-binaries.sh script so that it
removes the packages/validator/node_modules/@ibm-cloud and
packages/ruleset/node_modules/@ibm-cloud directories
prior to invoking pkg to build the executables.
This fixes a problem where the executables would end up with
the "old" versions of the openapi-ruleset and
openapi-ruleset-utilities dependencies in situations
where we're also publishing a new version of those dependencies
in the same build.
Removing the dependencies from packages/[validator,ruleset] means
that pkg would end up using the dependencies in the project's
top-level "node_modules" directory instead, which will have the correct
versions of the dependencies, whether or not we've published a new
version of them during the same build.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>