fix(ibm-requestbody-name): turn rule off by default, is now redundant (#673)
Now that the IBM API Handbook provides firm and robust guidance on schema
naming and we validate for this guidance in 'ibm-schema-naming-convention',
requiring the 'x-codegen-request-body-name' to be specified on relevant
schemas is no longer necessary. It is resulting in APIs having to duplicate
names between the schema name, title, and extension. This rule will remain
in place for teams that wish to use it but this commit turns it off by default.

https://cloud.ibm.com/docs/api-handbook?topic=api-handbook-schemas#naming

Signed-off-by: Dustin Popp <dpopp07@gmail.com>