fix: make schema path de-resolution utility more robust
Before, the check used faulty logic to determine if we were on the last
iteration through a loop. This could lead to unintended behavior in rare
cases. This clarifies the logic to only execute the last time through the
loop and for no other reason (the potential reason being if the last path
segment matched an earlier path segment).

Signed-off-by: Dustin Popp <dpopp07@gmail.com>