fix(ibm-collection-array-property): avoid crashing when spectral adjusts paths
I found some odd behavior in which Spectral messes with the path strings reported
in the "path" array in the rule's context. This led to our code crashing when we
couldn't find the string in the API. This change skips the rule and logs an error.

In the future, we may want to address the problem on the Spectral side or fix the
path in-line and proceed with the check.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>