feat: add utility for checking composite schemas with a looser constraint
The existing `schemaHasConstraint` function will only determine a schema meets
the given constraint when every element of a `oneOf` or `anyOf` list. This is
desired behavior in the majority of cases. However, in some cases, it is
helpful to note when a schema includes a `oneOf` or `anyOf` list with only a
single element meeting the constraint, instead of all of them. The new utility
function, `schemaLooselyHasConstraint`, delivers this functionality.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>