fix(ibm-well-defined-dictionaries): flag dictionaries with dictionary-type values
The API Handbook states that the values for dictionaries should not themselves
be dictionaries (no dictionaries of dictionaries). Previously, this rule would
not catch that scenario. Now, it will.

Additionally, more robust logic is used to check for the presence of ambiguous
dictionaries. The old logic would miss flagging valid violations in complex,
composed schemas without the fix.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>
Co-authored-by: Phil Adams <phil_adams@us.ibm.com>