chore: add checkbox for scoring rubric to pr template
Now that we have a rubric containing scoring data for all rules in our ruleset,
we need to ensure that each PR adding a new rule to the ruleset updates the
rubric with information for the new rule. This commit adds an item to the
checklist in our PR template.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>