build: separate docker deployment into tagged release build
This offers greater control to retry the Docker deployment step in the event
that it fails, independently of the release being published. We've observed
the network request being fragile during builds, so it is important to be
able to retry this step without re-running the release publishing step.

This commit also removes the Travis badge from the README and aligns the
naming and formatting in the build workflow with our other projects.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>