fix(ibm-api-symmetry): print info logs in coherent order
The code to determine if one schema is a graph fragment of the other
uses a depth-first algorithm that prints a log, if it determines the
schema to violate the graph fragment pattern, with the reason behind
the violation. Due to the depth-first nature of the algorithm, the
logs are currently printed in depth first order, which is not as
coherent for the user to read.

This change introduces a stack to collect the logs during processing
and print them in reverse order afterwards, to give the user a better
sense of what happened during the processing.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>