fix: ingore violations from invalid locations of an api definition
Some extensions in an API definition, like those providing SDK-related metadata,
are not valid places to check for rule violations.

Since some built-in, Spectral rules use broad JSON paths for their rules, it is
possible for them to report false positives within extension objects, even when
the extension objects contain information that is completely unrelated to the
rule (like the typed-enum rule).

To avoid this in a generic way, this change filters out any rule violations that
occur within our internal, SDK-related metadata extension and provides for any
potential locations that may be added in the future.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>