fix(ibm-use-date-based-format): tighten heuristic for flagging date-time values (#717)
The regular expressions used to determine if a value should be considered a
date-time value are too loose and allow false positives to sneak in. It's a
tricky problem, but this commit makes an attempt to tighten them by dismissing
the possibility of a date-time value if any letters are found that are not
relevant to a date-time value.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>