feat: add helpers to extract processed api from spectral context
The Spectral 'context' object contains processed data about the
OpenAPI definition being validated, including the resolved and
unresolved versions of the definition and information about
references. We use these objects in rule functions and they need
to be extracted from deeply nested fields in the spectral context.
Rather than always needing to remember or look up the fields,
this commit allows us to use simple helper functions to get the
data we need.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>