fix(ibm-valid-schema-example): support circular references in schemas
This rule would previously fail if it tried to validate the example for
a schema that contains a circular reference (i.e. defines a schema cycle)
because the ref resolver leaves these schemas with unresolved refs.
The JSON Schema validator tries to resolve the references but can't find
them, so it fails. This commit unconditionally adds the components from
the OpenAPI definition to the schema so that the JSON Schema parser can
find references if it needs to.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>